Alteração temp.py

+++ Primeiros-codigos/temp.py
@@ -1,3 +1,3 @@
-x = int(input('Informe a temperatura em Celsius: ', ))
+x = float(input('Informe a temperatura em Celsius: ', ))
 c = (x * 9/5)+32
 print('A temperatura em Fahrenheit: ' ,c,'°F')
